+++ packages/rum/src/boot/rum.spec.ts
@@ -4,37 +4,20 @@ import { RumPerformanceNavigationTiming } from '../browser/performanceCollection
 
 import { LifeCycle, LifeCycleEventType } from '../domain/lifeCycle'
 import { SESSION_KEEP_ALIVE_INTERVAL, THROTTLE_VIEW_UPDATE_PERIOD } from '../domain/rumEventsCollection/view/trackViews'
+import { RumEvent } from '../rumEvent.types'
 import { startRumEventCollection } from './rum'
 
-interface ServerRumEvent {
-  application_id: string
-  date: number
-  type: string
-  evt: {
-    category: string
-  }
-  session_id: string
-  session: {
-    id: string
-  }
-  view: {
-    id: string
-    referrer: string
-    url: string
-  }
-}
-
 function collectServerEvents(lifeCycle: LifeCycle) {
-  const serverRumEvents: ServerRumEvent[] = []
+  const serverRumEvents: RumEvent[] = []
   lifeCycle.subscribe(LifeCycleEventType.RUM_EVENT_COLLECTED, ({ serverRumEvent }) => {
-    serverRumEvents.push(serverRumEvent as any)
+    serverRumEvents.push(serverRumEvent)
   })
   return serverRumEvents
 }
 
 describe('rum session', () => {
   let setupBuilder: TestSetupBuilder
-  let serverRumEvents: ServerRumEvent[]
+  let serverRumEvents: RumEvent[]
 
   beforeEach(() => {
     if (isIE()) {
@@ -43,7 +26,10 @@ describe('rum session', () => {
 
     setupBuilder = setup().beforeBuild(({ applicationId, location, lifeCycle, configuration, session }) => {
       serverRumEvents = collectServerEvents(lifeCycle)
-      return startRumEventCollection(applicationId, location, lifeCycle, configuration, session, () => ({}))
+      return startRumEventCollection(applicationId, location, lifeCycle, configuration, session, () => ({
+        context: {},
+        user: {},
+      }))
     })
   })
 
@@ -80,7 +66,7 @@ describe('rum session', () => {
 describe('rum session keep alive', () => {
   let isSessionTracked: boolean
   let setupBuilder: TestSetupBuilder
-  let serverRumEvents: ServerRumEvent[]
+  let serverRumEvents: RumEvent[]
 
   beforeEach(() => {
     if (isIE()) {
@@ -96,7 +82,10 @@ describe('rum session keep alive', () => {
       })
       .beforeBuild(({ applicationId, location, lifeCycle, configuration, session }) => {
         serverRumEvents = collectServerEvents(lifeCycle)
-        return startRumEventCollection(applicationId, location, lifeCycle, configuration, session, () => ({}))
+        return startRumEventCollection(applicationId, location, lifeCycle, configuration, session, () => ({
+          context: {},
+          user: {},
+        }))
       })
   })
 
@@ -151,12 +140,15 @@ describe('rum view url', () => {
   const VIEW_DURATION = 1000
 
   let setupBuilder: TestSetupBuilder
-  let serverRumEvents: ServerRumEvent[]
+  let serverRumEvents: RumEvent[]
 
   beforeEach(() => {
     setupBuilder = setup().beforeBuild(({ applicationId, location, lifeCycle, configuration, session }) => {
       serverRumEvents = collectServerEvents(lifeCycle)
-      return startRumEventCollection(applicationId, location, lifeCycle, configuration, session, () => ({}))
+      return startRumEventCollection(applicationId, location, lifeCycle, configuration, session, () => ({
+        context: {},
+        user: {},
+      }))
     })
   })