+++ packages/rum/src/domain/rumEventsCollection/action/actionCollection.spec.ts
@@ -1,5 +1,5 @@
 import { setup, TestSetupBuilder } from '../../../../test/specHelper'
-import { RumEventType } from '../../../types'
+import { RumEventType } from '../../../rawRumEvent.types'
 import { LifeCycleEventType } from '../../lifeCycle'
 import { startActionCollection } from './actionCollection'
 import { ActionType } from './trackActions'