+++ packages/rum/src/domain/rumEventsCollection/view/viewCollection.spec.ts
@@ -1,5 +1,5 @@
 import { setup, TestSetupBuilder } from '../../../../test/specHelper'
-import { RumEventType } from '../../../types'
+import { RumEventType } from '../../../rawRumEvent.types'
 import { LifeCycleEventType } from '../../lifeCycle'
 import { View, ViewLoadingType } from './trackViews'
 import { startViewCollection } from './viewCollection'