+++ packages/rum/src/index.ts
@@ -1,1 +1,10 @@
 export { RumUserConfiguration, RumGlobal, datadogRum } from './boot/rum.entry'
+export {
+  RumEvent,
+  RumActionEvent,
+  CommonProperties,
+  RumErrorEvent,
+  RumViewEvent,
+  RumResourceEvent,
+  RumLongTaskEvent,
+} from './rumEvent.types'